TIL something fundamental about encryption
Encrypting the same string twice yields different hashes,
which breaks the way we were dealing with denormalising the
tags in the database.

That makes total sense though, those math people are pretty
clever, turns out.

So for now, simply not encrypting the tags as this is a
little PoC anyway.